backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
numpy
passlib[bcrypt]
pydantic
email-validator
